backend/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
SQLAlchemy>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
mysql-connector-python>=9.0.0
httpx>=0.25.0
pandas>=2.2.0
bleach>=6.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
prefect>=2.14.0
